Crate mc_rpc

Starknet RPC server API implementation

It uses the madara client and backend in order to answer queries.

Structs

Starknet

A Starknet RPC server for Madara

Traits

StarknetRpcApiServer

Server trait implementation for the StarknetRpcApi RPC API.

Functions

casm_contract_class_to_compiled_class

Converts a [CasmContractClass] to a [CompiledClass]

flattened_sierra_to_casm_contract_class

Converts a [FlattenedSierraClass] to a [CasmContractClass]

get_block_by_block_hash

Returns the current Starknet block from the block header’s digest

get_casm_cotract_class_hash

to_declare_transaction

to_rpc_contract_class

Returns a [ContractClass] from a [BlockifierContractClass]

to_tx

Converts a broadcasted transaction to a transaction Supports Invoke, Declare and DeployAccount transactions