Trait mc_rpc::StarknetRpcApiServer

pub trait StarknetRpcApiServer: Sized + Send + Sync + 'static {

    // Required methods
    fn block_number(&self) -> Result<u64, Error>;
    fn block_hash_and_number(&self) -> Result<BlockHashAndNumber, Error>;
    fn get_block_transaction_count(
        &self,
        block_id: BlockId
    ) -> Result<u128, Error>;
    fn get_storage_at(
        &self,
        contract_address: FieldElement,
        key: FieldElement,
        block_id: BlockId
    ) -> Result<Felt, Error>;
    fn call(
        &self,
        request: FunctionCall,
        block_id: BlockId
    ) -> Result<Vec<String, Global>, Error>;
    fn get_class_at(
        &self,
        block_id: BlockId,
        contract_address: FieldElement
    ) -> Result<ContractClass, Error>;
    fn get_class_hash_at(
        &self,
        block_id: BlockId,
        contract_address: FieldElement
    ) -> Result<Felt, Error>;
    fn syncing<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<SyncStatusType, Error>> + Send + 'async_trait, Global>>
       where 'life0: 'async_trait,
             Self: 'async_trait;
    fn get_class(
        &self,
        block_id: BlockId,
        class_hash: FieldElement
    ) -> Result<ContractClass, Error>;
    fn get_block_with_tx_hashes(
        &self,
        block_id: BlockId
    ) -> Result<MaybePendingBlockWithTxHashes, Error>;
    fn get_nonce(
        &self,
        block_id: BlockId,
        contract_address: FieldElement
    ) -> Result<Felt, Error>;
    fn get_block_with_txs(
        &self,
        block_id: BlockId
    ) -> Result<MaybePendingBlockWithTxs, Error>;
    fn chain_id(&self) -> Result<Felt, Error>;
    fn add_invoke_transaction<'life0, 'async_trait>(
        &'life0 self,
        invoke_transaction: BroadcastedInvokeTransaction
    ) -> Pin<Box<dyn Future<Output = Result<InvokeTransactionResult, Error>> + Send + 'async_trait, Global>>
       where 'life0: 'async_trait,
             Self: 'async_trait;
    fn add_deploy_account_transaction<'life0, 'async_trait>(
        &'life0 self,
        deploy_account_transaction: BroadcastedDeployAccountTransaction
    ) -> Pin<Box<dyn Future<Output = Result<DeployAccountTransactionResult, Error>> + Send + 'async_trait, Global>>
       where 'life0: 'async_trait,
             Self: 'async_trait;
    fn estimate_fee<'life0, 'async_trait>(
        &'life0 self,
        request: BroadcastedTransaction,
        block_id: BlockId
    ) -> Pin<Box<dyn Future<Output = Result<FeeEstimate, Error>> + Send + 'async_trait, Global>>
       where 'life0: 'async_trait,
             Self: 'async_trait;
    fn get_transaction_by_block_id_and_index(
        &self,
        block_id: BlockId,
        index: usize
    ) -> Result<Transaction, Error>;
    fn get_state_update(&self, block_id: BlockId) -> Result<StateUpdate, Error>;
    fn pending_transactions<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<Vec<Transaction, Global>, Error>> + Send + 'async_trait, Global>>
       where 'life0: 'async_trait,
             Self: 'async_trait;
    fn get_events<'life0, 'async_trait>(
        &'life0 self,
        filter: EventFilterWithPage
    ) -> Pin<Box<dyn Future<Output = Result<EventsPage, Error>> + Send + 'async_trait, Global>>
       where 'life0: 'async_trait,
             Self: 'async_trait;
    fn add_declare_transaction<'life0, 'async_trait>(
        &'life0 self,
        declare_transaction: BroadcastedDeclareTransaction
    ) -> Pin<Box<dyn Future<Output = Result<DeclareTransactionResult, Error>> + Send + 'async_trait, Global>>
       where 'life0: 'async_trait,
             Self: 'async_trait;
    fn get_transaction_by_hash(
        &self,
        transaction_hash: FieldElement
    ) -> Result<Transaction, Error>;
    fn get_transaction_receipt(
        &self,
        transaction_hash: FieldElement
    ) -> Result<MaybePendingTransactionReceipt, Error>;

    // Provided method
    fn into_rpc(self) -> RpcModule<Self> { ... }

}

Server trait implementation for the StarknetRpcApi RPC API.

Required Methods

fn block_number(&self) -> Result<u64, Error>

Get the most recent accepted block number

fn block_hash_and_number(&self) -> Result<BlockHashAndNumber, Error>

Get the most recent accepted block hash and number

fn get_block_transaction_count(&self, block_id: BlockId) -> Result<u128, Error>

Get the number of transactions in a block given a block id

fn get_storage_at( &self, contract_address: FieldElement, key: FieldElement, block_id: BlockId ) -> Result<Felt, Error>

Get the value of the storage at the given address and key, at the given block id

fn call( &self, request: FunctionCall, block_id: BlockId ) -> Result<Vec<String, Global>, Error>

Call a contract function at a given block id

fn get_class_at( &self, block_id: BlockId, contract_address: FieldElement ) -> Result<ContractClass, Error>

Get the contract class at a given contract address for a given block id

fn get_class_hash_at( &self, block_id: BlockId, contract_address: FieldElement ) -> Result<Felt, Error>

Get the contract class hash in the given block for the contract deployed at the given address

fn syncing<'life0, 'async_trait>( &'life0 self ) -> Pin<Box<dyn Future<Output = Result<SyncStatusType, Error>> + Send + 'async_trait, Global>>where 'life0: 'async_trait, Self: 'async_trait,

Get an object about the sync status, or false if the node is not syncing

fn get_class( &self, block_id: BlockId, class_hash: FieldElement ) -> Result<ContractClass, Error>

Get the contract class definition in the given block associated with the given hash

fn get_block_with_tx_hashes( &self, block_id: BlockId ) -> Result<MaybePendingBlockWithTxHashes, Error>

Get block information with transaction hashes given the block id

fn get_nonce( &self, block_id: BlockId, contract_address: FieldElement ) -> Result<Felt, Error>

Get the nonce associated with the given address at the given block

fn get_block_with_txs( &self, block_id: BlockId ) -> Result<MaybePendingBlockWithTxs, Error>

Get block information with full transactions given the block id

fn chain_id(&self) -> Result<Felt, Error>

Get the chain id

fn add_invoke_transaction<'life0, 'async_trait>( &'life0 self, invoke_transaction: BroadcastedInvokeTransaction ) -> Pin<Box<dyn Future<Output = Result<InvokeTransactionResult, Error>> + Send + 'async_trait, Global>>where 'life0: 'async_trait, Self: 'async_trait,

Add an Invoke Transaction to invoke a contract function

fn add_deploy_account_transaction<'life0, 'async_trait>( &'life0 self, deploy_account_transaction: BroadcastedDeployAccountTransaction ) -> Pin<Box<dyn Future<Output = Result<DeployAccountTransactionResult, Error>> + Send + 'async_trait, Global>>where 'life0: 'async_trait, Self: 'async_trait,

Add a Deploy Account Transaction

fn estimate_fee<'life0, 'async_trait>( &'life0 self, request: BroadcastedTransaction, block_id: BlockId ) -> Pin<Box<dyn Future<Output = Result<FeeEstimate, Error>> + Send + 'async_trait, Global>>where 'life0: 'async_trait, Self: 'async_trait,

Estimate the fee associated with transaction

fn get_transaction_by_block_id_and_index( &self, block_id: BlockId, index: usize ) -> Result<Transaction, Error>

Get the details of a transaction by a given block id and index

fn get_state_update(&self, block_id: BlockId) -> Result<StateUpdate, Error>

Get the information about the result of executing the requested block

fn pending_transactions<'life0, 'async_trait>( &'life0 self ) -> Pin<Box<dyn Future<Output = Result<Vec<Transaction, Global>, Error>> + Send + 'async_trait, Global>>where 'life0: 'async_trait, Self: 'async_trait,

Returns the transactions in the transaction pool, recognized by this sequencer

fn get_events<'life0, 'async_trait>( &'life0 self, filter: EventFilterWithPage ) -> Pin<Box<dyn Future<Output = Result<EventsPage, Error>> + Send + 'async_trait, Global>>where 'life0: 'async_trait, Self: 'async_trait,

Returns all events matching the given filter

fn add_declare_transaction<'life0, 'async_trait>( &'life0 self, declare_transaction: BroadcastedDeclareTransaction ) -> Pin<Box<dyn Future<Output = Result<DeclareTransactionResult, Error>> + Send + 'async_trait, Global>>where 'life0: 'async_trait, Self: 'async_trait,

Submit a new transaction to be added to the chain

fn get_transaction_by_hash( &self, transaction_hash: FieldElement ) -> Result<Transaction, Error>

Returns the information about a transaction by transaction hash.

fn get_transaction_receipt( &self, transaction_hash: FieldElement ) -> Result<MaybePendingTransactionReceipt, Error>

Returns the receipt of a transaction by transaction hash.

Provided Methods

fn into_rpc(self) -> RpcModule<Self>

Collects all the methods and subscriptions defined in the trait and adds them into a single RpcModule.